requirements.txt
# Core dependencies for GPU memorization evaluation
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Data processing and utilities
numpy>=1.24.0
matplotlib>=3.6.0
tqdm>=4.64.0
pyyaml>=6.0